[FE] REFACTOR: 우측 구역 공통 컴포넌트로 이동 #710
