+++ frontend_v4/src/Layout.tsx
@@ -1,20 +1,22 @@
 import React, { useState, useEffect } from "react";
 import { Outlet } from "react-router";
 import { useNavigate, useLocation } from "react-router-dom";
-import { useSetRecoilState } from "recoil";
-import { userState } from "@/recoil/atoms";
+import { useRecoilValue, useSetRecoilState } from "recoil";
+import { toggleCabinetInfoState, userState } from "@/recoil/atoms";
 import TopNav from "@/components/TopNav";
-import LeftNavContainer from "@/containers/LeftNavContainer";
-import LeftNavOptionContainer from "@/containers/LeftNavOptionContainer";
+import LeftNavAreaContainer from "./containers/LeftNavAreaContainer";
 import LoadingModal from "@/components/LoadingModal";
 import { getCookie } from "@/api/react_cookie/cookies";
 import { axiosMyInfo } from "@/api/axios/axios.custom";
 import { UserDto } from "@/types/dto/user.dto";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
+import CabinetInfoArea from "./components/CabinetInfoArea";
+import MapInfoContainer from "./containers/MapInfoContainer";
 
 const Layout = (): JSX.Element => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [isValidToken, setIsValidToken] = useState<boolean>(false);
+  const toggleCabinetInfo = useRecoilValue(toggleCabinetInfoState);
   const setUser = useSetRecoilState<UserDto>(userState);
   const navigate = useNavigate();
   const location = useLocation();
@@ -52,9 +54,18 @@ const Layout = (): JSX.Element => {
         <LoadingModal />
       ) : (
         <WrapperStyled>
-          <LeftNavContainer />
-          <LeftNavOptionContainer isVisible={!isHomePage} />
-          <Outlet />
+          <LeftNavAreaContainer isVisible={!isHomePage} />
+          <MainStyled>
+            <Outlet />
+          </MainStyled>
+          <DetailInfoContainerStyled
+            id="cabinetDetailArea"
+            className={toggleCabinetInfo ? "on" : ""}
+            isHomePage={isHomePage}
+          >
+            <CabinetInfoArea />
+          </DetailInfoContainerStyled>
+          <MapInfoContainer />
         </WrapperStyled>
       )}
     </React.Fragment>
@@ -68,4 +79,32 @@ const WrapperStyled = styled.div`
   height: 100%;
   display: flex;
   overflow: hidden;
+`;
+
+const MainStyled = styled.main`
+  width: 100%;
+  height: 100%;
+  overflow-x: hidden;
+`;
+
+const DetailInfoContainerStyled = styled.div<{ isHomePage: boolean }>`
+  min-width: 330px;
+  padding-top: 45px;
+  border-left: 1px solid var(--line-color);
+  background-color: var(--white);
+  ${(props) =>
+    props.isHomePage &&
+    css`
+      position: fixed;
+      top: 80px;
+      right: 0;
+      height: calc(100% - 80px);
+      z-index: 9;
+      transform: translateX(120%);
+      transition: transform 0.3s ease-in-out;
+      box-shadow: 0 0 40px 0 var(--bg-shadow);
+      &.on {
+        transform: translateX(0%);
+      }
+    `}
 `;
