+++ frontend_v4/src/containers/LeftNavContainer.tsx
@@ -11,14 +11,13 @@ import {
   currentFloorNumberState,
   currentFloorCabinetState,
   currentSectionNameState,
-  toggleMapInfoState,
-  toggleCabinetInfoState,
 } from "@/recoil/atoms";
 import { currentLocationFloorState } from "@/recoil/selectors";
 import { axiosCabinetByLocationFloor } from "@/api/axios/axios.custom";
 import { CabinetInfoByLocationFloorDto } from "@/types/dto/cabinet.dto";
 import useLeftNav from "@/hooks/useLeftNav";
 import { removeCookie } from "@/api/react_cookie/cookies";
+import useDetailInfo from "@/hooks/useDetailInfo";
 
 const LeftNavContainer = () => {
   const floors = useRecoilValue<Array<number>>(currentLocationFloorState);
@@ -30,11 +29,10 @@ const LeftNavContainer = () => {
     CabinetInfoByLocationFloorDto[]
   >(currentFloorCabinetState);
   const setCurrentSection = useSetRecoilState<string>(currentSectionNameState);
-  const toggleMapInfo = useSetRecoilState(toggleMapInfoState);
-  const toggleCabinetInfo = useSetRecoilState(toggleCabinetInfoState);
   const navigator = useNavigate();
   const { pathname } = useLocation();
   const { closeLeftNav } = useLeftNav();
+  const { closeDetailInfo } = useDetailInfo();
 
   useEffect(() => {
     if (currentFloor === undefined) return;
@@ -58,8 +56,7 @@ const LeftNavContainer = () => {
   };
 
   const onClickHomeButton = () => {
-    toggleMapInfo(() => false);
-    toggleCabinetInfo(() => false);
+    closeDetailInfo();
     resetCurrentFloor();
     closeLeftNav();
     navigator("/home");
