+++ frontend_v4/src/containers/SectionPaginationContainer.tsx
@@ -48,7 +48,7 @@ export default SectionPaginationContainer;
 
 const SectionPaginationStyled = styled.div`
   min-width: 360px;
-  margin-bottom: 30px;
+  margin: 30px 0px;
   display: flex;
   flex-direction: column;
   align-items: center;
